+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Examples/Cast.hs
@@ -61,7 +61,6 @@ import Cardano.Ledger.PoolParams (
   PoolMetadata (..),
   PoolParams (..),
  )
-import Cardano.Ledger.Slot (SlotNo (..))
 import Cardano.Protocol.Crypto (hashVerKeyVRF)
 import Cardano.Protocol.TPraos.OCert (KESPeriod (..))
 import qualified Data.ByteString.Char8 as BS (pack)
@@ -120,7 +119,7 @@ aliceSHK = (KeyHashObj . hashKey . vKey) aliceStake
 
 -- | Alice's base address
 alicePtrAddr :: Addr
-alicePtrAddr = Addr Testnet alicePHK (StakeRefPtr $ Ptr (SlotNo 10) minBound minBound)
+alicePtrAddr = Addr Testnet alicePHK (StakeRefPtr $ Ptr 10 minBound minBound)
 
 -- | Alice's stake pool parameters
 alicePoolParams :: PoolParams