+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/TerseTools.hs
@@ -16,7 +16,6 @@ import Cardano.Ledger.Credential (
 import Cardano.Ledger.Shelley.LedgerState (IncrementalStake (..))
 import Cardano.Ledger.TxIn (TxId (..), TxIn (..))
 import Cardano.Ledger.UTxO (UTxO (..))
-import Cardano.Slotting.Slot (SlotNo (..))
 import qualified Data.Map.Strict as Map
 
 -- ====================================================
@@ -85,7 +84,7 @@ instance Terse (Credential r) where
   terse (KeyHashObj (KeyHash hash)) = "Key " ++ show hash
 
 instance Terse Ptr where
-  terse (Ptr (SlotNo n) i j) = "Ptr " ++ show n ++ " " ++ show i ++ " " ++ show j
+  terse (Ptr n i j) = "Ptr " ++ show n ++ " " ++ show i ++ " " ++ show j
 
 instance Terse TxId where
   terse (TxId safehash) = show (extractHash safehash)