+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Examples/MirTransfer.hs
@@ -8,6 +8,7 @@ where
 
 import Cardano.Ledger.BaseTypes (Mismatch (..), ProtVer (..), natVersion)
 import Cardano.Ledger.Coin (Coin (..), DeltaCoin (..))
+import Cardano.Ledger.Credential (SlotNo32 (..))
 import Cardano.Ledger.Hashes (GenDelegs (..))
 import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.API (
@@ -49,12 +50,13 @@ env pv acnt =
   DelegEnv
     { slotNo = slot
     , deCurEpochNo = epochFromSlotNo slot
-    , ptr_ = Ptr slot minBound minBound
+    , ptr_ = Ptr slot32 minBound minBound
     , acnt_ = acnt
     , ppDE = emptyPParams & ppProtocolVersionL .~ pv
     }
   where
     slot = SlotNo 50
+    slot32 = SlotNo32 50
 
 shelleyPV :: ProtVer
 shelleyPV = ProtVer (natVersion @2) 0