+++ libs/cardano-ledger-core/src/Cardano/Ledger/Plutus/TxInfo.hs
@@ -48,11 +48,10 @@ import Cardano.Crypto.Hash.Class (hashToBytes)
 import Cardano.Ledger.Address (Addr (..), RewardAccount (..))
 import Cardano.Ledger.BaseTypes (
   BoundedRational (unboundRational),
+  CertIx (..),
   EpochInterval (..),
   EpochNo (..),
-  TxIx,
-  certIxToInt,
-  txIxToInt,
+  TxIx (..),
  )
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders (
@@ -65,7 +64,7 @@ import Cardano.Ledger.Binary.Coders (
  )
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Core
-import Cardano.Ledger.Credential (Credential (..), Ptr (..), StakeReference (..))
+import Cardano.Ledger.Credential (Credential (..), Ptr (..), SlotNo32 (..), StakeReference (..))
 import Cardano.Ledger.Plutus.Data (Data (..), getPlutusData)
 import Cardano.Ledger.Plutus.ExUnits (ExUnits (..))
 import Cardano.Ledger.TxIn (TxId (..), TxIn (..), txInToText)
@@ -134,10 +133,8 @@ transScriptHash (ScriptHash h) = PV1.ScriptHash (PV1.toBuiltin (hashToBytes h))
 
 transStakeReference :: StakeReference -> Maybe PV1.StakingCredential
 transStakeReference (StakeRefBase cred) = Just (PV1.StakingHash (transCred cred))
-transStakeReference (StakeRefPtr (Ptr (SlotNo slot) txIx certIx)) =
-  let !txIxInteger = toInteger (txIxToInt txIx)
-      !certIxInteger = toInteger (certIxToInt certIx)
-   in Just (PV1.StakingPtr (fromIntegral slot) txIxInteger certIxInteger)
+transStakeReference (StakeRefPtr (Ptr (SlotNo32 slot) (TxIx txIx) (CertIx certIx))) =
+  Just (PV1.StakingPtr (toInteger slot) (toInteger txIx) (toInteger certIx))
 transStakeReference StakeRefNull = Nothing
 
 transCred :: Credential kr -> PV1.Credential
@@ -177,7 +174,7 @@ transTxId :: TxId -> PV1.TxId
 transTxId (TxId safe) = PV1.TxId (transSafeHash safe)
 
 transTxIn :: TxIn -> PV1.TxOutRef
-transTxIn (TxIn txid txIx) = PV1.TxOutRef (transTxId txid) (toInteger (txIxToInt txIx))
+transTxIn (TxIn txid (TxIx txIx)) = PV1.TxOutRef (transTxId txid) (toInteger txIx)
 
 transCoinToValue :: Coin -> PV1.Value
 transCoinToValue (Coin c) = PV1.singleton PV1.adaSymbol PV1.adaToken c