+++ libs/cardano-ledger-core/test/Test/Cardano/Ledger/AddressSpec.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-deprecations #-}
 
 module Test.Cardano.Ledger.AddressSpec (spec) where