+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/TreeDiff.hs
@@ -167,6 +167,8 @@ instance ToExpr (Credential keyrole)
 
 instance ToExpr StakeReference
 
+deriving newtype instance ToExpr SlotNo32
+
 instance ToExpr Ptr
 
 deriving newtype instance