+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/Instances/Ledger.hs
@@ -583,6 +583,9 @@ instance (IsConwayUniv fn, Typeable b) => HasSpec fn (AbstractHash Blake2b_224 b
 instance HasSimpleRep StakeReference
 instance IsConwayUniv fn => HasSpec fn StakeReference
 
+instance HasSimpleRep SlotNo32
+instance IsConwayUniv fn => HasSpec fn SlotNo32
+
 instance HasSimpleRep Ptr
 instance IsConwayUniv fn => HasSpec fn Ptr