+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.17.0.0
 
+* Add `SlotNo32` and use it in `Ptr` definition
+* Add `mkPtrNormalized`
+* Deprecate `normalizePtr` and `addrPtrNormalize`
 * Switch `TxIx` and `CertIx` to use `Word16` instead of `Word64`
 * Deprecate `mkTxIx` and `mkCertIx`
 * Extract `indexProxy` from `HashAnnotated` type class and deprecate it.
@@ -54,6 +57,7 @@
 
 ### `testlib`
 
+* Deprecate `genBadPtr`, `genAddrBadPtr` and `genCompactAddrBadPtr`
 * Remove crypto parametrization from types: `KeyPair` and `KeyPairs`
 
 ## 1.16.0.0